@@ -1,66 +1,152 @@
 {
-  "name": "node-fetch",
-  "version": "2.6.0",
-  "description": "A light-weight module that brings window.fetch to node.js",
-  "main": "lib/index",
-  "browser": "./browser.js",
-  "module": "lib/index.mjs",
-  "files": [
-    "lib/index.js",
-    "lib/index.mjs",
-    "lib/index.es.js",
-    "browser.js"
-  ],
-  "engines": {
-    "node": "4.x || >=6.0.0"
-  },
-  "scripts": {
-    "build": "cross-env BABEL_ENV=rollup rollup -c",
-    "prepare": "npm run build",
-    "test": "cross-env BABEL_ENV=test mocha --require babel-register --throw-deprecation test/test.js",
-    "report": "cross-env BABEL_ENV=coverage nyc --reporter lcov --reporter text mocha -R spec test/test.js",
-    "coverage": "cross-env BABEL_ENV=coverage nyc --reporter json --reporter text mocha -R spec test/test.js && codecov -f coverage/coverage-final.json"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/bitinn/node-fetch.git"
-  },
-  "keywords": [
-    "fetch",
-    "http",
-    "promise"
-  ],
-  "author": "David Frank",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/bitinn/node-fetch/issues"
-  },
-  "homepage": "https://github.com/bitinn/node-fetch",
-  "devDependencies": {
-    "@ungap/url-search-params": "^0.1.2",
-    "abort-controller": "^1.1.0",
-    "abortcontroller-polyfill": "^1.3.0",
-    "babel-core": "^6.26.3",
-    "babel-plugin-istanbul": "^4.1.6",
-    "babel-preset-env": "^1.6.1",
-    "babel-register": "^6.16.3",
-    "chai": "^3.5.0",
-    "chai-as-promised": "^7.1.1",
-    "chai-iterator": "^1.1.1",
-    "chai-string": "~1.3.0",
-    "codecov": "^3.3.0",
-    "cross-env": "^5.2.0",
-    "form-data": "^2.3.3",
-    "is-builtin-module": "^1.0.0",
-    "mocha": "^5.0.0",
-    "nyc": "11.9.0",
-    "parted": "^0.1.1",
-    "promise": "^8.0.3",
-    "resumer": "0.0.0",
-    "rollup": "^0.63.4",
-    "rollup-plugin-babel": "^3.0.7",
-    "string-to-arraybuffer": "^1.0.2",
-    "whatwg-url": "^5.0.0"
-  },
-  "dependencies": {}
+    "name": "node-fetch",
+    "version": "3.0.0-beta.1",
+    "description": "A light-weight module that brings window.fetch to node.js",
+    "main": "dist/index.js",
+    "module": "dist/index.mjs",
+    "types": "types/index.d.ts",
+    "files": [
+        "src/**/*",
+        "dist/**/*",
+        "types/**/*.d.ts"
+    ],
+    "engines": {
+        "node": ">=10.0.0"
+    },
+    "scripts": {
+        "build": "pika-pack --out dist/",
+        "prepare": "npm run build",
+        "prepublishOnly": "npm run build",
+        "test": "cross-env BABEL_ENV=test mocha --require @babel/register --throw-deprecation test/*.js",
+        "report": "cross-env BABEL_ENV=coverage nyc --reporter lcov --reporter text mocha -R spec test/*.js",
+        "coverage": "cross-env BABEL_ENV=coverage nyc --reporter json --reporter text mocha -R spec test/*.js && codecov -f coverage/coverage-final.json",
+        "lint": "xo"
+    },
+    "repository": {
+        "type": "git",
+        "url": "https://github.com/node-fetch/node-fetch.git"
+    },
+    "keywords": [
+        "fetch",
+        "http",
+        "promise"
+    ],
+    "author": "David Frank",
+    "license": "MIT",
+    "bugs": {
+        "url": "https://github.com/node-fetch/node-fetch/issues"
+    },
+    "homepage": "https://github.com/node-fetch/node-fetch",
+    "funding": {
+        "type": "opencollective",
+        "url": "https://opencollective.com/node-fetch"
+    },
+    "devDependencies": {
+        "@babel/core": "^7.8.7",
+        "@babel/preset-env": "^7.8.7",
+        "@babel/register": "^7.8.6",
+        "@pika/pack": "^0.5.0",
+        "@pika/plugin-build-node": "^0.9.2",
+        "@pika/plugin-build-types": "^0.9.2",
+        "@pika/plugin-copy-assets": "^0.9.2",
+        "@pika/plugin-standard-pkg": "^0.9.2",
+        "abort-controller": "^3.0.0",
+        "abortcontroller-polyfill": "^1.4.0",
+        "chai": "^4.2.0",
+        "chai-as-promised": "^7.1.1",
+        "chai-iterator": "^3.0.2",
+        "chai-string": "^1.5.0",
+        "codecov": "^3.6.5",
+        "cross-env": "^7.0.2",
+        "form-data": "^3.0.0",
+        "mocha": "^7.1.0",
+        "nyc": "^15.0.0",
+        "parted": "^0.1.1",
+        "promise": "^8.1.0",
+        "resumer": "0.0.0",
+        "string-to-arraybuffer": "^1.0.2",
+        "xo": "^0.28.0"
+    },
+    "dependencies": {
+        "data-uri-to-buffer": "^3.0.0",
+        "fetch-blob": "^1.0.5"
+    },
+    "@pika/pack": {
+        "pipeline": [
+            [
+                "@pika/plugin-standard-pkg"
+            ],
+            [
+                "@pika/plugin-build-node"
+            ],
+            [
+                "@pika/plugin-build-types"
+            ],
+            [
+                "@pika/plugin-copy-assets",
+                {
+                    "files": [
+                        "externals.d.ts"
+                    ]
+                }
+            ]
+        ]
+    },
+    "xo": {
+        "envs": [
+            "node",
+            "browser"
+        ],
+        "rules": {
+            "complexity": 0,
+            "promise/prefer-await-to-then": 0,
+            "no-mixed-operators": 0,
+            "no-negated-condition": 0,
+            "unicorn/prevent-abbreviations": 0
+        },
+        "ignores": [
+            "dist"
+        ],
+        "overrides": [
+            {
+                "files": "test/**/*.js",
+                "envs": [
+                    "node",
+                    "mocha"
+                ],
+                "rules": {
+                    "max-nested-callbacks": 0,
+                    "no-unused-expressions": 0,
+                    "new-cap": 0,
+                    "guard-for-in": 0
+                }
+            },
+            {
+                "files": "example.js",
+                "rules": {
+                    "import/no-extraneous-dependencies": 0
+                }
+            }
+        ]
+    },
+    "babel": {
+        "presets": [
+            [
+                "@babel/preset-env",
+                {
+                    "targets": {
+                        "node": true
+                    }
+                }
+            ]
+        ]
+    },
+    "nyc": {
+        "require": [
+            "@babel/register"
+        ],
+        "sourceMap": false,
+        "instrument": false
+    },
+    "runkitExampleFilename": "example.js"
 }
