@@ -1,24 +1,19 @@
 import * as http from 'http';
-import { parse } from 'url';
 import * as zlib from 'zlib';
-import * as stream from 'stream';
-import { multipart as Multipart } from 'parted';
-
-let convert;
-try { convert = require('encoding').convert; } catch(e) {}
+import {multipart as Multipart} from 'parted';
 
 export default class TestServer {
 	constructor() {
 		this.server = http.createServer(this.router);
 		this.port = 30001;
 		this.hostname = 'localhost';
-		// node 8 default keepalive timeout is 5000ms
+		// Node 8 default keepalive timeout is 5000ms
 		// make it shorter here as we want to close server quickly at the end of tests
 		this.server.keepAliveTimeout = 1000;
-		this.server.on('error', function(err) {
+		this.server.on('error', err => {
 			console.log(err.stack);
 		});
-		this.server.on('connection', function(socket) {
+		this.server.on('connection', socket => {
 			socket.setTimeout(1500);
 		});
 	}
@@ -31,8 +26,22 @@ export default class TestServer {
 		this.server.close(cb);
 	}
 
-	router(req, res) {
-		let p = parse(req.url).pathname;
+	mockResponse(responseHandler) {
+		this.server.nextResponseHandler = responseHandler;
+		return `http://${this.hostname}:${this.port}/mocked`;
+	}
+
+	router(request, res) {
+		const p = request.url;
+
+		if (p === '/mocked') {
+			if (this.nextResponseHandler) {
+				this.nextResponseHandler(res);
+				this.nextResponseHandler = undefined;
+			} else {
+				throw new Error('No mocked response. Use ’TestServer.mockResponse()’.');
+			}
+		}
 
 		if (p === '/hello') {
 			res.statusCode = 200;
@@ -70,7 +79,11 @@ export default class TestServer {
 			res.statusCode = 200;
 			res.setHeader('Content-Type', 'text/plain');
 			res.setHeader('Content-Encoding', 'gzip');
-			zlib.gzip('hello world', function(err, buffer) {
+			zlib.gzip('hello world', (err, buffer) => {
+				if (err) {
+					throw err;
+				}
+
 				res.end(buffer);
 			});
 		}
@@ -79,9 +92,26 @@ export default class TestServer {
 			res.statusCode = 200;
 			res.setHeader('Content-Type', 'text/plain');
 			res.setHeader('Content-Encoding', 'gzip');
-			zlib.gzip('hello world', function(err, buffer) {
-				// truncate the CRC checksum and size check at the end of the stream
-				res.end(buffer.slice(0, buffer.length - 8));
+			zlib.gzip('hello world', (err, buffer) => {
+				if (err) {
+					throw err;
+				}
+
+				// Truncate the CRC checksum and size check at the end of the stream
+				res.end(buffer.slice(0, -8));
+			});
+		}
+
+		if (p === '/gzip-capital') {
+			res.statusCode = 200;
+			res.setHeader('Content-Type', 'text/plain');
+			res.setHeader('Content-Encoding', 'GZip');
+			zlib.gzip('hello world', (err, buffer) => {
+				if (err) {
+					throw err;
+				}
+
+				res.end(buffer);
 			});
 		}
 
@@ -89,7 +119,11 @@ export default class TestServer {
 			res.statusCode = 200;
 			res.setHeader('Content-Type', 'text/plain');
 			res.setHeader('Content-Encoding', 'deflate');
-			zlib.deflate('hello world', function(err, buffer) {
+			zlib.deflate('hello world', (err, buffer) => {
+				if (err) {
+					throw err;
+				}
+
 				res.end(buffer);
 			});
 		}
@@ -99,18 +133,25 @@ export default class TestServer {
 			res.setHeader('Content-Type', 'text/plain');
 			if (typeof zlib.createBrotliDecompress === 'function') {
 				res.setHeader('Content-Encoding', 'br');
-				zlib.brotliCompress('hello world', function (err, buffer) {
+				zlib.brotliCompress('hello world', (err, buffer) => {
+					if (err) {
+						throw err;
+					}
+
 					res.end(buffer);
 				});
 			}
 		}
 
-
 		if (p === '/deflate-raw') {
 			res.statusCode = 200;
 			res.setHeader('Content-Type', 'text/plain');
 			res.setHeader('Content-Encoding', 'deflate');
-			zlib.deflateRaw('hello world', function(err, buffer) {
+			zlib.deflateRaw('hello world', (err, buffer) => {
+				if (err) {
+					throw err;
+				}
+
 				res.end(buffer);
 			});
 		}
@@ -130,7 +171,7 @@ export default class TestServer {
 		}
 
 		if (p === '/timeout') {
-			setTimeout(function() {
+			setTimeout(() => {
 				res.statusCode = 200;
 				res.setHeader('Content-Type', 'text/plain');
 				res.end('text');
@@ -141,7 +182,7 @@ export default class TestServer {
 			res.statusCode = 200;
 			res.setHeader('Content-Type', 'text/plain');
 			res.write('test');
-			setTimeout(function() {
+			setTimeout(() => {
 				res.end('test');
 			}, 1000);
 		}
@@ -155,10 +196,10 @@ export default class TestServer {
 		if (p === '/size/chunk') {
 			res.statusCode = 200;
 			res.setHeader('Content-Type', 'text/plain');
-			setTimeout(function() {
+			setTimeout(() => {
 				res.write('test');
 			}, 10);
-			setTimeout(function() {
+			setTimeout(() => {
 				res.end('test');
 			}, 20);
 		}
@@ -167,69 +208,6 @@ export default class TestServer {
 			res.statusCode = 200;
 			res.setHeader('Content-Type', 'text/plain');
 			res.end('testtest');
-		}
-
-		if (p === '/encoding/gbk') {
-			res.statusCode = 200;
-			res.setHeader('Content-Type', 'text/html');
-			res.end(convert('<meta charset="gbk"><div>中文</div>', 'gbk'));
-		}
-
-		if (p === '/encoding/gb2312') {
-			res.statusCode = 200;
-			res.setHeader('Content-Type', 'text/html');
-			res.end(convert('<meta http-equiv="Content-Type" content="text/html; charset=gb2312"><div>中文</div>', 'gb2312'));
-		}
-
-		if (p === '/encoding/gb2312-reverse') {
-			res.statusCode = 200;
-			res.setHeader('Content-Type', 'text/html');
-			res.end(convert('<meta content="text/html; charset=gb2312" http-equiv="Content-Type"><div>中文</div>', 'gb2312'));
-		}
-
-		if (p === '/encoding/shift-jis') {
-			res.statusCode = 200;
-			res.setHeader('Content-Type', 'text/html; charset=Shift-JIS');
-			res.end(convert('<div>日本語</div>', 'Shift_JIS'));
-		}
-
-		if (p === '/encoding/euc-jp') {
-			res.statusCode = 200;
-			res.setHeader('Content-Type', 'text/xml');
-			res.end(convert('<?xml version="1.0" encoding="EUC-JP"?><title>日本語</title>', 'EUC-JP'));
-		}
-
-		if (p === '/encoding/utf8') {
-			res.statusCode = 200;
-			res.end('中文');
-		}
-
-		if (p === '/encoding/order1') {
-			res.statusCode = 200;
-			res.setHeader('Content-Type', 'charset=gbk; text/plain');
-			res.end(convert('中文', 'gbk'));
-		}
-
-		if (p === '/encoding/order2') {
-			res.statusCode = 200;
-			res.setHeader('Content-Type', 'text/plain; charset=gbk; qs=1');
-			res.end(convert('中文', 'gbk'));
-		}
-
-		if (p === '/encoding/chunked') {
-			res.statusCode = 200;
-			res.setHeader('Content-Type', 'text/html');
-			res.setHeader('Transfer-Encoding', 'chunked');
-			res.write('a'.repeat(10));
-			res.end(convert('<meta http-equiv="Content-Type" content="text/html; charset=Shift_JIS" /><div>日本語</div>', 'Shift_JIS'));
-		}
-
-		if (p === '/encoding/invalid') {
-			res.statusCode = 200;
-			res.setHeader('Content-Type', 'text/html');
-			res.setHeader('Transfer-Encoding', 'chunked');
-			res.write('a'.repeat(1200));
-			res.end(convert('中文', 'gbk'));
 		}
 
 		if (p === '/redirect/301') {
@@ -276,7 +254,7 @@ export default class TestServer {
 		if (p === '/redirect/slow') {
 			res.statusCode = 301;
 			res.setHeader('Location', '/redirect/301');
-			setTimeout(function() {
+			setTimeout(() => {
 				res.end();
 			}, 1000);
 		}
@@ -284,7 +262,7 @@ export default class TestServer {
 		if (p === '/redirect/slow-chain') {
 			res.statusCode = 301;
 			res.setHeader('Location', '/redirect/slow');
-			setTimeout(function() {
+			setTimeout(() => {
 				res.end();
 			}, 10);
 		}
@@ -355,12 +333,14 @@ export default class TestServer {
 			res.statusCode = 200;
 			res.setHeader('Content-Type', 'application/json');
 			let body = '';
-			req.on('data', function(c) { body += c });
-			req.on('end', function() {
+			request.on('data', c => {
+				body += c;
+			});
+			request.on('end', () => {
 				res.end(JSON.stringify({
-					method: req.method,
-					url: req.url,
-					headers: req.headers,
+					method: request.method,
+					url: request.url,
+					headers: request.headers,
 					body
 				}));
 			});
@@ -369,26 +349,32 @@ export default class TestServer {
 		if (p === '/multipart') {
 			res.statusCode = 200;
 			res.setHeader('Content-Type', 'application/json');
-			const parser = new Multipart(req.headers['content-type']);
+			const parser = new Multipart(request.headers['content-type']);
 			let body = '';
-			parser.on('part', function(field, part) {
+			parser.on('part', (field, part) => {
 				body += field + '=' + part;
 			});
-			parser.on('end', function() {
+			parser.on('end', () => {
 				res.end(JSON.stringify({
-					method: req.method,
-					url: req.url,
-					headers: req.headers,
-					body: body
+					method: request.method,
+					url: request.url,
+					headers: request.headers,
+					body
 				}));
 			});
-			req.pipe(parser);
+			request.pipe(parser);
+		}
+
+		if (p === '/m%C3%B6bius') {
+			res.statusCode = 200;
+			res.setHeader('Content-Type', 'text/plain');
+			res.end('ok');
 		}
 	}
 }
 
 if (require.main === module) {
-	const server = new TestServer;
+	const server = new TestServer();
 	server.start(() => {
 		console.log(`Server started listening at port ${server.port}`);
 	});
