@@ -1,45 +1,53 @@
 
 /**
- * request.js
+ * Request.js
  *
  * Request class contains server only options
  *
  * All spec algorithm step numbers are based on https://fetch.spec.whatwg.org/commit-snapshots/ae716822cb3a61843226cd090eefc6589446c1d2/.
  */
 
-import Url from 'url';
+import {format as formatUrl} from 'url';
 import Stream from 'stream';
-import Headers, { exportNodeCompatibleHeaders } from './headers.js';
-import Body, { clone, extractContentType, getTotalBytes } from './body';
+import Headers, {exportNodeCompatibleHeaders} from './headers';
+import Body, {clone, extractContentType, getTotalBytes} from './body';
+import {isAbortSignal} from './utils/is';
 
 const INTERNALS = Symbol('Request internals');
-
-// fix an issue where "format", "parse" aren't a named export for node <10
-const parse_url = Url.parse;
-const format_url = Url.format;
 
 const streamDestructionSupported = 'destroy' in Stream.Readable.prototype;
 
 /**
- * Check if a value is an instance of Request.
+ * Check if `obj` is an instance of Request.
  *
- * @param   Mixed   input
- * @return  Boolean
+ * @param  {*} obj
+ * @return {boolean}
  */
-function isRequest(input) {
+function isRequest(object) {
 	return (
-		typeof input === 'object' &&
-		typeof input[INTERNALS] === 'object'
+		typeof object === 'object' &&
+		typeof object[INTERNALS] === 'object'
 	);
 }
 
-function isAbortSignal(signal) {
-	const proto = (
-		signal
-		&& typeof signal === 'object'
-		&& Object.getPrototypeOf(signal)
-	);
-	return !!(proto && proto.constructor.name === 'AbortSignal');
+/**
+ * Wrapper around `new URL` to handle relative URLs (https://github.com/nodejs/node/issues/12682)
+ *
+ * @param  {string} urlStr
+ * @return {void}
+ */
+function parseURL(urlString) {
+	/*
+		Check whether the URL is absolute or not
+
+		Scheme: https://tools.ietf.org/html/rfc3986#section-3.1
+		Absolute URL: https://tools.ietf.org/html/rfc3986#section-4.3
+	*/
+	if (/^[a-zA-Z][a-zA-Z\d+\-.]*:/.exec(urlString)) {
+		return new URL(urlString);
+	}
+
+	throw new TypeError('Only absolute URLs are supported');
 }
 
 /**
@@ -53,35 +61,38 @@ export default class Request {
 	constructor(input, init = {}) {
 		let parsedURL;
 
-		// normalize input
+		// Normalize input and force URL to be encoded as UTF-8 (https://github.com/bitinn/node-fetch/issues/245)
 		if (!isRequest(input)) {
 			if (input && input.href) {
-				// in order to support Node.js' Url objects; though WHATWG's URL objects
+				// In order to support Node.js' Url objects; though WHATWG's URL objects
 				// will fall into this branch also (since their `toString()` will return
 				// `href` property anyway)
-				parsedURL = parse_url(input.href);
+				parsedURL = parseURL(input.href);
 			} else {
-				// coerce input to a string before attempting to parse
-				parsedURL = parse_url(`${input}`);
+				// Coerce input to a string before attempting to parse
+				parsedURL = parseURL(`${input}`);
 			}
+
 			input = {};
 		} else {
-			parsedURL = parse_url(input.url);
+			parsedURL = parseURL(input.url);
 		}
 
 		let method = init.method || input.method || 'GET';
 		method = method.toUpperCase();
 
+		// eslint-disable-next-line no-eq-null, eqeqeq
 		if ((init.body != null || isRequest(input) && input.body !== null) &&
 			(method === 'GET' || method === 'HEAD')) {
 			throw new TypeError('Request with GET/HEAD method cannot have body');
 		}
 
-		let inputBody = init.body != null ?
+		// eslint-disable-next-line no-eq-null, eqeqeq
+		const inputBody = init.body != null ?
 			init.body :
-			isRequest(input) && input.body !== null ?
+			(isRequest(input) && input.body !== null ?
 				clone(input) :
-				null;
+				null);
 
 		Body.call(this, inputBody, {
 			timeout: init.timeout || input.timeout || 0,
@@ -90,19 +101,21 @@ export default class Request {
 
 		const headers = new Headers(init.headers || input.headers || {});
 
-		if (inputBody != null && !headers.has('Content-Type')) {
+		if (inputBody !== null && !headers.has('Content-Type')) {
 			const contentType = extractContentType(inputBody);
 			if (contentType) {
 				headers.append('Content-Type', contentType);
 			}
 		}
 
-		let signal = isRequest(input)
-			? input.signal
-			: null;
-		if ('signal' in init) signal = init.signal
+		let signal = isRequest(input) ?
+			input.signal :
+			null;
+		if ('signal' in init) {
+			signal = init.signal;
+		}
 
-		if (signal != null && !isAbortSignal(signal)) {
+		if (signal !== null && !isAbortSignal(signal)) {
 			throw new TypeError('Expected signal to be an instanceof AbortSignal');
 		}
 
@@ -111,18 +124,19 @@ export default class Request {
 			redirect: init.redirect || input.redirect || 'follow',
 			headers,
 			parsedURL,
-			signal,
+			signal
 		};
 
-		// node-fetch-only options
+		// Node-fetch-only options
 		this.follow = init.follow !== undefined ?
-			init.follow : input.follow !== undefined ?
-			input.follow : 20;
+			init.follow : (input.follow !== undefined ?
+				input.follow : 20);
 		this.compress = init.compress !== undefined ?
-			init.compress : input.compress !== undefined ?
-			input.compress : true;
+			init.compress : (input.compress !== undefined ?
+				input.compress : true);
 		this.counter = init.counter || input.counter || 0;
 		this.agent = init.agent || input.agent;
+		this.highWaterMark = init.highWaterMark || input.highWaterMark;
 	}
 
 	get method() {
@@ -130,7 +144,7 @@ export default class Request {
 	}
 
 	get url() {
-		return format_url(this[INTERNALS].parsedURL);
+		return formatUrl(this[INTERNALS].parsedURL);
 	}
 
 	get headers() {
@@ -165,12 +179,12 @@ Object.defineProperty(Request.prototype, Symbol.toStringTag, {
 });
 
 Object.defineProperties(Request.prototype, {
-	method: { enumerable: true },
-	url: { enumerable: true },
-	headers: { enumerable: true },
-	redirect: { enumerable: true },
-	clone: { enumerable: true },
-	signal: { enumerable: true },
+	method: {enumerable: true},
+	url: {enumerable: true},
+	headers: {enumerable: true},
+	redirect: {enumerable: true},
+	clone: {enumerable: true},
+	signal: {enumerable: true}
 });
 
 /**
@@ -180,10 +194,10 @@ Object.defineProperties(Request.prototype, {
  * @return  Object   The options object to be passed to http.request
  */
 export function getNodeRequestOptions(request) {
-	const parsedURL = request[INTERNALS].parsedURL;
+	const {parsedURL} = request[INTERNALS];
 	const headers = new Headers(request[INTERNALS].headers);
 
-	// fetch step 1.3
+	// Fetch step 1.3
 	if (!headers.has('Accept')) {
 		headers.set('Accept', '*/*');
 	}
@@ -198,24 +212,26 @@ export function getNodeRequestOptions(request) {
 	}
 
 	if (
-		request.signal
-		&& request.body instanceof Stream.Readable
-		&& !streamDestructionSupported
+		request.signal &&
+		request.body instanceof Stream.Readable &&
+		!streamDestructionSupported
 	) {
-		throw new Error('Cancellation of streamed requests with AbortSignal is not supported in node < 8');
+		throw new Error('Cancellation of streamed requests with AbortSignal is not supported');
 	}
 
 	// HTTP-network-or-cache fetch steps 2.4-2.7
 	let contentLengthValue = null;
-	if (request.body == null && /^(POST|PUT)$/i.test(request.method)) {
+	if (request.body === null && /^(post|put)$/i.test(request.method)) {
 		contentLengthValue = '0';
 	}
-	if (request.body != null) {
+
+	if (request.body !== null) {
 		const totalBytes = getTotalBytes(request);
 		if (typeof totalBytes === 'number') {
 			contentLengthValue = String(totalBytes);
 		}
 	}
+
 	if (contentLengthValue) {
 		headers.set('Content-Length', contentLengthValue);
 	}
@@ -230,7 +246,7 @@ export function getNodeRequestOptions(request) {
 		headers.set('Accept-Encoding', 'gzip,deflate');
 	}
 
-	let agent = request.agent;
+	let {agent} = request;
 	if (typeof agent === 'function') {
 		agent = agent(parsedURL);
 	}
@@ -242,9 +258,21 @@ export function getNodeRequestOptions(request) {
 	// HTTP-network fetch step 4.2
 	// chunked encoding is handled by Node.js
 
-	return Object.assign({}, parsedURL, {
+	// manually spread the URL object instead of spread syntax
+	const requestOptions = {
+		path: parsedURL.pathname,
+		pathname: parsedURL.pathname,
+		hostname: parsedURL.hostname,
+		protocol: parsedURL.protocol,
+		port: parsedURL.port,
+		hash: parsedURL.hash,
+		search: parsedURL.search,
+		query: parsedURL.query,
+		href: parsedURL.href,
 		method: request.method,
 		headers: exportNodeCompatibleHeaders(headers),
 		agent
-	});
+	};
+
+	return requestOptions;
 }
